% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gapminder.R
\docType{data}
\name{my_gapminder}
\alias{my_gapminder}
\title{Jennifer Bryan's gapminder Data}
\format{a data frame with 1704 observations and 6 variables:
\describe{
\item{country}{country name}
\item{continent}{continent name}
\item{year}{from 1952 to 2007}
\item{lifeExp}{life expectancy at birth}
\item{pop}{total population}
\item{gdpPercap}{per-capita GDP}
}

\url{https://github.com/jennybc/gapminder}

The data were collected by Jennifer Bryan.}
\usage{
my_gapminder
}
\description{
The famous \code{\link[gapminder]{gapminder}} dataset taken from the
\code{gapminder} package to demonstrate use of all of these functions.
}
\keyword{datasets}
