% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Input data frame}

\item{cl}{True class value of trainning data}

\item{k_nn}{Integer representing the number of neighbors}

\item{k_cv}{Integer representing the number of folds}
}
\value{
class:  a vector of the predicted class for all observations

cv_err: a numeric with the cross-validation misclassification error
}
\description{
This function performs k-Nearest Neighbors Cross-Validation.
}
\examples{
data(my_iris)
my_knn_cv(iris[,-5],iris$Species,5,5)

}
\keyword{prediction}
