% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear model function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula that will be used for fitting the \code{data} set.}

\item{data}{A given data set that will be used for the model.}
}
\value{
a summary table of coefficients

standard error

t value and p value for a hypothesis test
}
\description{
This function fits a linear model in R
}
\examples{
data(my_gapminder)
my_lm(lifeExp ~ gdpPercap + continent, data = my_gapminder)

}
\keyword{inference}
