% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t_test}
\alias{my_t_test}
\title{T.test function}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{A numeric vector of data.}

\item{alternative}{A character string specifying the alternative hypothesis.
This should only accept "two.sided", "less", or "greater". Otherwise, throw an informative error.}

\item{mu}{A number indicating the null hypothesis value of the mean.}
}
\value{
test_stat: the numeric test statistic.

df: the degrees of freedom.

\code{alternative}: the value of the parameter \code{alternative}.

p_val: the numeric p-value.
}
\description{
This function performs a one sample t-test.
}
\examples{
my_t_test(rnorm(10, mean = 1.5, sd = 1), alternative = "greater", mu = 0)
my_t_test(rnorm(10, mean = 1.5, sd = 1), alternative = "less", mu = 0)
my_t_test(rnorm(10, mean = 1.5, sd = 1), alternative = "two.sided", mu = 0)

}
\keyword{inference}
